move Neo to Neo folder; added DayCell
