+++ src/Plugin.Maui.Calendar/Shared/Controls/Neo/DayCell.xaml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Border
+    x:Class="Plugin.Maui.Calendar.Controls.DayCell"
+    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:controls="clr-namespace:RaptorCalendar.Controls"
+    x:Name="DayBorder"
+    x:DataType="controls:DayCell"
+    BackgroundColor="White"
+    HeightRequest="30"
+    HorizontalOptions="Center"
+    Loaded="DayCell_Loaded"
+    StrokeShape="RoundRectangle 3"
+    StrokeThickness="1"
+    VerticalOptions="Center"
+    WidthRequest="30">
+    <Border.Resources>
+        <ResourceDictionary>
+            <Style x:Key="DaysLabelStyle" TargetType="Label">
+                <Setter Property="HorizontalTextAlignment" Value="Center" />
+                <Setter Property="TextColor" Value="{StaticResource TextGray}" />
+                <Setter Property="FontSize" Value="14" />
+                <Setter Property="LineBreakMode" Value="WordWrap" />
+                <Setter Property="VerticalTextAlignment" Value="Center" />
+                <Setter Property="Margin" Value="5,2" />
+            </Style>
+        </ResourceDictionary>
+    </Border.Resources>
+    <Label
+        x:Name="DayLabel"
+        HorizontalOptions="Center"
+        Style="{StaticResource DaysLabelStyle}"
+        TextColor="Black"
+        VerticalOptions="Center" />
+    <Border.GestureRecognizers>
+        <TapGestureRecognizer Tapped="Handle_Tapped" />
+    </Border.GestureRecognizers>
+</Border>
