+++ src/Plugin.Maui.Calendar/Shared/Controls/Neo/DefaultFooterSectionNeo.xaml.cs
