+++ src/Plugin.Maui.Calendar/Shared/Controls/Neo/DefaultHeaderSectionNeo.xaml.cs
